fix: test shared level hash calculator
